refactor: move get delegations to delegations controller (#164)
