+++ connector-api/src/main/java/it/pagopa/selfcare/mscore/api/DelegationConnector.java
@@ -8,5 +8,5 @@ public interface DelegationConnector {
 
     Delegation save(Delegation delegation);
 
-    List<Delegation> find(String from, String productId);
+    List<Delegation> find(String from, String to, String productId);
 }
